+++ HISTORY.rst
@@ -12,6 +12,7 @@ HEAD (unreleased)
 - More prefetching for large list for speedups.
 - Fixing organism/reference editing.
 - More refined checking of barcodes and demultiplexing cycles.
+- Properly catching error of duplicate barcode name/sequence in barcode set.
 
 ------
 v1.0.0